Merge `date.js`.
Merged `date.js` into the main library. Caveats when talking about
library size are demoralizing. With `date.js` merged, the size of the
library is what it is. No extensions. No bait and switch.

In documentation, I'm discussion extension through functional
composition, not through plug-ins and what not. A plug-in implies that
I've left functionality out of the library in order to make claims about
the minified size of the library. The argument that it is up to the user
to decide sounds disingenuous.

Thus, I've taken to considering the size of the library to be the size
of `timezone.js` and `date.js` combined. With this commit I've combined
them.

Combining them, of course, makes them smaller than summing the sizes of
the two files. It gives `gzip` more repetition to chew on.

Combined the files are under 3K.

    File                             Raw         Min         Zip     Min/Zip
    lib/timezone.js          15574/15.21  7931/07.75  4187/04.09  3030/02.96