Use `timezone` as `require` root.
Zone and locale files are now in the root directory of the packaged
Timezone library published to NPM. This means that an application can
require a locale using `require("timezone/pl_PL")`, or require timezone
data using `require("timezone/America/Detroit")`.

This is less verbose than `require("timezone/zones/America/Detroit")`,
but still not as terse as `require("zones/America/Detroit")`. The
problem with using `zones` is that `locales` is not available. If it
were available, it would be ambiguous implying that it was providing
general local information for more than Timezone.

I'd considered using `strftime` as a package name to store locales, but
that is taken, of course, by a JavaScript implementation of `strftime`.

I've just now discovered that `z` is available. I'm content to leave the
locales in the timezone package, meaning that I could use it to create
`require("z/America/Detroit")`.

There is a problem with separating the zone information into a separate
package. I'd no longer be able to say that Timezone has no dependencies
without a disclaimer. It is essentially true, but as it stands it is
absolutely true.

Then you'll have a problem of versioning. If I make a change to the
format, that will mean that the `z` package will not work with older
libraries.

Plus, the build would become even more complicated.

I've tried to consider what makes this question so prickly for me.
Something about naming is hard to settle. This seems like it is
something that concerns me a lot because at some point, the method of
initialization for Timezone seemed controversial to me. It seems much
less so the further I get into documentation. Users will either load all
the zones at once, or else load them individually. In the former case,
that is going to be `require("timezone/zones")`, in the latter case it
will be a formatted require string.

The annoyance of loading won't go away.