Accept any numeric object as POSIX time.
If the first argument can be converted to a number, it is used as the
date. The numberic value is floored to remove any fractional part. This
is as much clean up as we can do. Someone can still give us
1.7976931348623157e+308 and the output of our date formats will consist
mostly of `NaN`.

With this change, we can now accept `new Date()` as an input value.