Remove `parseInt`.
The implicit conversion from string to integer does not treat leading
zeros as an indicator of octal. I was using to `parseInt` so I could
ensure that the conversions were decimal. Using implicit conversions
reduces the size of the code.

Added some tests for padding. I was altering tests to add leading zeros
to the optional padding width, so it seemed like a good time to create
some tests that focused on padding specifiers.

    File                             Raw         Min         Zip     Min/Zip
    timezone.js              14969/14.62  7535/07.36  4084/03.99  2901/02.83