Retry unknown string parameters.
When invoking the **Timezone** function with a rule set that has not
already been loaded and the name of a time zone defined by the rule set,
if the rules come after the name, the time zone rules are not applied.
This is because the name is valid if a rule set can be found in the
context object, but the context object has not yet been assigned the
rules because we work through the parameter list left to right.

With this commit, we create a `retry` array and push it onto the end of
our variable arguments array. If we don't understand a string parameter,
we push it onto the `retry` array.

Arrays on the variable arguments array are flattened into the variable
arguments array (if they are not the initial date array argument.) The
retry array will be flattened into place and any strings we could not
understand will be retries.

If we still don't understand them, we push them onto the `retry` array,
but the `retry` array has already been flattened into the variable
arguments array, so it is no longer in the variable arguments array to
be flattened again.

Added an assertion to `t/offset/load.t` to ensure that this continues to
work.

Renamed `splat` to `vargs` in `convert`.

    File                             Raw         Min         Zip     Min/Zip
    timezone.js              15098/14.74  7578/07.40  4114/04.02  2915/02.85