Date arrays begin with `"@"`.
To reduce ambiguity between an array of zone or locale definitions and a
date array, the first element of a date array is now the character
`"@"`, the sigil for a Perl array, because I once spent a lot of time
with Perl.

This reduction in ambiguity means that the **Timezone** function will
flatten any argument array. An argument list can be built in an array
and given directly to the **Timezone** function, without having to
invoke `Function.apply`.

Given that the crux of the **Timezone** API is specified is the
arguments array, this optional invocation method will probably prove to
be more than a nice to have.