Thorough validation.
This is thorough validation of arguments, adding over a kilobyte in size
to the library. It now teeters on the brink of 3k.

What this represents is a departure from the simplicity of the one
function interface, to something of a janky domain-specific language.
The introduction of the `"@"` to the date array to reduce ambiguity is
actually the introduction of types. The exceptions are type checking.

If that is what you want, then make a `TimeZoneConverter` object, name
the functions, and throw all kinds of exceptions.

The whole thing is off in the wrong direction, but I'm going to create a
release, so I can look at might have been. Also, it mapped out all the
different failure paths, so I'd like to continue to build off that.

    File                             Raw         Min         Zip     Min/Zip
    timezone.js              15320/14.96  7944/07.76  4214/04.12  3050/02.98