Fail silently, JavaScript date array.
Backed out the `"@"` at the start of a date array.

Adopted the JavaScript convention of a zero indexed month, because
everything else about the library is about conventions. It's easier to
bemoan the JavaScript caveat that to document our caveat to the
JavaScript caveat.

Removed all the extra exception throwing and validity checking. Back to
the old rules. If we don't know what it is, we ignore it. If you confuse
us, we are confused. If it hurts when do this, don't do that.

The majority of the time we will be called like so...

    tz(Date.now(), "Europe/Paris", "%c");

The time zone will have to be selected from a list of time zones. People
are not going to key in a time zone, requiring validation. If validation
is required, it can easily be implemented.

    if (tz(0, "Europe/Paris", "%Z") == "UTC") {
      throw new Error("Europe/Paris is not loaded.");
    }

When we cannot convert to POSIX time because the time does not exist in
the time zone, we return `null` instead of raising an exception. Up to
the application developer check for it.

    File                             Raw         Min         Zip     Min/Zip
    timezone.js              14631/14.29  7488/07.31  4028/03.93  2880/02.81