Use `eggert/tz` GitHub repo as database source.
Now using the `eggert/tz` GitHub repository as the source for the
database, instead of downloading and unraveling the IANA TimeZone
database tarballs. `eggert/tz` promises to be the de facto unofficial TZ
repository, seeing as how Eggert is one of the traditional maintainers
of the Olson files.

See [unofficial TZ repository on
GitHub](http://mm.icann.org/pipermail/tz/2012-July/018126.html) for more
details on the move from a private SCCS repository to `git` with GitHub
hosting.

This Timezone module will always endeavour to point the `vendor/tz`
submodule to the commit on `eggert/tz` that mirrors the latest official
release.