Add `.js` suffix to test programs.
Proof now uses `.t.js` as its conventional file suffix because the Ruby
gem that GitHub uses to determine the programming language a source file
&mdash; Linguist &mdash; will always decide that a `.t` file is a Perl
test file.

The `.js` suffix generally meets the expectations of most tools. The
`.t` suffix looked nice in the reporting, but it wasn't so dear.

One of our tests is written in Ruby. It uses the suffix `*.t.rb`.